// packages/typespec-ts/src/rlc-common/interfaces.ts
export interface RLCModel {
  libraryName: string;
  srcPath: string;
  paths: Paths;
  importInfo: ImportInfo;
  options?: RLCOptions;
  schemas: Schema[];
  apiVersionInfo?: ApiVersionInfo;
  parameters?: OperationParameter[];
  responses?: OperationResponse[];
  helperDetails?: HelperFunctionDetails;
  urlInfo?: UrlInfo;
  telemetryOptions?: TelemetryInfo;
  sampleGroups?: RLCSampleGroup[];
  rlcSourceDir?: string;
}

export interface ImportInfo {
  internalImports: Imports;
  runtimeImports: Imports;
}

export type Imports = Record<ImportType, ImportMetadata>;
// ...
export interface RLCSampleGroup {
  filename: string;
  clientPackageName: string;
  defaultFactoryName: string;
  samples: RLCSampleDetail[];
  importedTypes?: string[];
}

/**
 * An independent sample detail and it will be wrapped as a func
 */
export interface RLCSampleDetail {
  /**
   * metadata for comments
   */
  description: string;
  originalFileLocation?: string;
  name: string;
  path: string;
  defaultFactoryName: string;
  clientParamAssignments: string[];
  pathParamAssignments: string[];
  methodParamAssignments: string[];
  clientParamNames: string;
  pathParamNames: string;
  methodParamNames: "options" | "" | string;
  method: string;
  isLRO: boolean;
  isPaging: boolean;
  useLegacyLro: boolean;
}

export interface TelemetryInfo {
  customRequestIdHeaderName?: string;
}

export interface PathTemplateApiVersion {
  value: string;
  templateName: string;
}

export interface UrlInfo {
  endpoint?: string;
  urlParameters?: PathParameter[];
  apiVersionInfo?: ApiVersionInfo;
}

export interface ApiVersionInfo {
  definedPosition?: ApiVersionPosition;
  defaultValue?: string;
  isCrossedVersion?: boolean;
  required?: boolean;
}

export type ApiVersionPosition =
  | "path"
  | "query"
  | "baseurl"
  | "duplicate"
  | "none";
export interface HelperFunctionDetails {
  hasPaging?: boolean;
  hasLongRunning?: boolean;
  clientLroOverload?: boolean;
  pageDetails?: PagingDetails;
  hasMultiCollection?: boolean;
  hasPipeCollection?: boolean;
  hasSsvCollection?: boolean;
  hasTsvCollection?: boolean;
  hasCsvCollection?: boolean;
}

export interface PagingDetails {
  itemNames: string[];
  nextLinkNames: string[];
  isComplexPaging: boolean;
}
// ...
export interface PathMetadata {
  name: string;
  pathParameters: PathParameter[];
  methods: Methods;
  operationGroupName: string;
  description: string;
}

export type Paths = Record<string, PathMetadata>;

export type PathParameter = {
  oriName?: string;
  name: string;
  documentName?: string;
  type: string;
  description?: string;
  value?: string | number | boolean;
  wrapperType?: Schema;
};
// ...
export type PackageFlavor = "azure" | undefined;

export interface RLCOptions {
  /**
   * Whether to include response headers in the generated response types. If true, the generated response types will include headers as properties.
   */
  includeHeadersInResponse?: boolean;
  includeShortcuts?: boolean;
  multiClient?: boolean;
  batch?: any[];
  packageDetails?: PackageDetails;
  addCredentials?: boolean;
  /** Three possiblie values:
   * - undefined, no credentialScopes and relevant settings would be generated
   * - [], which means we would generate TokenCredential but no credentialScopes and relevant settings
   * - ["..."], which means we would generate credentialScopes and relevant settings with the given values
   */
  credentialScopes?: string[];
  credentialKeyHeaderName?: string;
  customHttpAuthHeaderName?: string;
  customHttpAuthSharedKeyPrefix?: string;
  /**
   * Three possible values:
   * - undefined, the default behavior which means we would generate metadata if the package.json file is absent
   * - true, which means we would always generate new files or override existing files
   * - false, which means we would not generate any files no matter there exists or not
   */
  generateMetadata?: boolean;
  /**
   * Three possible values:
   * - undefined, the default behavior which means we would generate test if there is no `test` folder
   * - true, which means we would always generate new files or override existing files
   * - false, which means we would not generate any files no matter there exists or not
   */
  generateTest?: boolean;
  generateSample?: boolean;
  azureSdkForJs?: boolean;
  azureOutputDirectory?: string;
  isTypeSpecTest?: boolean;
  title?: string;
  dependencyInfo?: DependencyInfo;
  productDocLink?: string;
  serviceInfo?: ServiceInfo;
  azureArm?: boolean;
  sourceFrom?: "TypeSpec" | "Swagger";
  isModularLibrary?: boolean;
  moduleKind?: "esm" | "cjs";
  enableOperationGroup?: boolean;
  flavor?: PackageFlavor;
  enableModelNamespace?: boolean;
  hierarchyClient?: boolean;
  compatibilityMode?: boolean;
  experimentalExtensibleEnums?: boolean;
  clearOutputFolder?: boolean;
  ignorePropertyNameNormalize?: boolean;
  ignoreEnumMemberNameNormalize?: boolean;
  compatibilityQueryMultiFormat?: boolean;
  typespecTitleMap?: Record<string, string>;
  hasSubscriptionId?: boolean;
  compatibilityLro?: boolean;
  ignoreNullableOnOptional?: boolean;
  isMultiService?: boolean;
  /**
   * When set to true, non-model return types (arrays, scalars, enums, bytes with binary content type)
   * will be wrapped in an XxxResponse type to maintain backward compatibility with HLC.
   * This option defaults to true for Azure flavor and to false otherwise, unless explicitly set.
   */
  wrapNonModelReturn?: boolean;
  /**
   * When enabled, every regular (non-LRO, non-paging) operation return type is augmented with a
   * `_response` property containing `rawResponse`, `parsedBody`, and `parsedHeaders`.
   */
  enableStorageCompat?: boolean;
  /**
   * When set to true, TypeSpec `unknown` type will be translated to `Record<string, unknown>`
   * instead of `any` in the generated Modular SDK. This is useful when migrating from HLC
   * where `unknown` in swagger mapped to `Record<string, unknown>`.
   */
  treatUnknownAsRecord?: boolean;
  /**
   * When set to true, generates React Native build targets (tsconfig, warp target, package.json exports).
   * Defaults to false. Only applicable when azureSdkForJs is true.
   */
  generateReactNativeTarget?: boolean;
  /**
   * When set to true, operations that can return either a body or a 204 empty body will have a
   * return type of `model | void` and include a body guard in the deserializer.
   * When false (default), the return type is just `model` and no body guard is emitted.
   */
  returnEmptyBody?: boolean;
}

export interface ServiceInfo {
  title?: string;
  description?: string;
}

export interface DependencyInfo {
  link: string;
  description: string;
}

export interface File {
  path: string;
  content: string;
}

export enum SchemaContext {
  /** Schema is used as an input to an operation. */
  Input = "input",
  /** Schema is used as an output from an operation. */
  Output = "output",
  /** Schema is used as an exception from an operation. */
  Exception = "exception"
}

export interface Schema {
  name: string;
  type: string;
  typeName?: string;
  outputTypeName?: string;
  description?: string;
  required?: boolean;
  default?: any;
  readOnly?: boolean;
  usage?: SchemaContext[];
  alias?: string;
  outputAlias?: string;
  fromCore?: boolean;
  enum?: any[];
  isConstant?: boolean;
}
// ...
export interface PackageDetails {
  name: string;
  scopeName?: string;
  nameWithoutScope?: string;
  description?: string;
  version?: string;
  isVersionUserProvided?: boolean;
}
export interface OperationParameter {
  operationGroup: string;
  operationName: string;
  /**
   * An operation with multiple request parameters means that
   * the operation can get different values for content-type and each value
   * may have a different type associated to it.
   */
  parameters: ParameterMetadatas[];
}

export interface ParameterMetadatas {
  parameters?: ParameterMetadata[];
  body?: ParameterBodyMetadata;
}
// ...
export interface OperationResponse {
  operationGroup: string;
  operationName: string;
  path: string;
  responses: ResponseMetadata[];
  // Check if the default response is one of superset of non-default responses
  isDefaultSupersetOfOthers?: boolean;
}
export interface ResponseMetadata {
  statusCode: string;
  description?: string;
  headers?: ResponseHeaderSchema[];
  body?: ResponseBodySchema;
  predefinedName?: string;
}